# shared/schemas/api_spec.yaml
openapi: 3.0.3
info:
  title: Club Project API
  description: API specification for Club Project microservices
  version: 1.0.0
  contact:
    name: Club Project Team
    email: team@club-project.com
servers:
  - url: https://api.club-project.com
    description: Production server
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /images/{path:path}:
    get:
      summary: Image pipeline proxy
      description: Proxy GET requests to image pipeline service
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response from image pipeline service
        '500':
          description: Internal server error

    post:
      summary: Image pipeline proxy
      description: Proxy POST requests to image pipeline service
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response from image pipeline service
        '400':
          description: Bad request
        '500':
          description: Internal server error
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...